requirements.txt
# Google API dependencies
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0

# GitHub API
PyGithub==2.1.1

# Template engine
Jinja2==3.1.3

# Configuration
PyYAML==6.0.1

# Secure credential storage (optional)
keyring==24.3.0

# Date/time handling
python-dateutil==2.8.2
pytz==2024.1
